// Source Code/Front End/src/app/shared/models/doctor/doctor-resp-data.ts
export class Doctor {
    doctorId: number;
    name: string;
    address1: string;
    address2: string;
    city: string;
    phone: string;
    mobile: string;
    email: string;
    salary: number;
    designation: string;
    gender: string;
    

    constructor() {
        this.doctorId=0;
        this.name = '';
        this.address1 = '';
        this.address2 = '';
        this.city = '';
        this.phone = '';
        this.mobile = '';
        this.email = '';
        this.salary = 0.0;
        this.designation = '';
        this.gender = '';
    }
}